fix: separate code into `widget-validation-helper.ts`
